Electrostatics 1.0.1

+++ src/main/java/com/vitaliirohozhyn_arsenisialitski/caecs/ecs/components/ChargeComponent.java
@@ -3,9 +3,9 @@ package com.vitaliirohozhyn_arsenisialitski.caecs.ecs.components;
 import com.vitaliirohozhyn_arsenisialitski.caecs.ecs.Component;
 
 public class ChargeComponent implements Component {
-    public Integer charge; // in Culombs(subject to change)
+    public Double charge; // in Culombs(subject to change)
 
-    public ChargeComponent(int a_charge) {
+    public ChargeComponent(double a_charge) {
         this.charge = a_charge;
     }
